pyproject.toml
[tool.poetry]
name = "poetry-new-project"
version = "0.1.0"
description = ""
authors = ["Smirnova Tanya <75753895+SmirnovaT@users.noreply.github.com>"]
readme = "README.md"
#packages = [{include = "/"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.1"
psycopg2 = "^2.9.5"
uvicorn = "^0.20.0"
fastapi = "^0.89.1"
alembic = "^1.9.2"
asyncpg = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
